Make RequestAuthorizer generic over its Token
Unfortunately, making `FinchProvider` generic over `Token` removes the
ability to use `FinchProvider` as an existential type (see the provider
registry for an example). `FinchProvider` now inherits from the internal
`_CallbackHandler` protocol, which can be used as an existential, to
work around this issue.

Additionally, because `identifier` is static (i.e., associated with the
type itself rather than any dynamic _instance_ of that type), it's not
possible for `AnyFinchProvider` to fulfill that requirement (as it
dynamically represents any possible provider type). Thus, we also
inherit from the internal `_FinchProvider` protocol, which doesn't
include the static identifier requirement.

Ultimately this allows us to make `RequestAuthorizer` generic over its
`Token` type, giving providers the flexibility to use whatever concrete
token type is appropriate.