Thread-safe pending requests during authentication
Introdces a `Channel` for one-to-many broadcast of the result of
authentication, allowing for thread-safe enqueueing of pending requests.

Introduces `Atomic` for expressive, closure-based scoping of critical
sections.

This is a dramatic overhaul of `RequestAuthorizer` internals, with
clearly documented critical sections and extracting useful concurrency
primitives.

`Channel` combines a `DispatchSemaphore` and `DispatchGroup` to allow
for thread-safe one-to-many broadcast of a message to interested
subscribers. In this case, subscribers are pending requests during
authentication. When authentication is complete, the result is published
to those pending requests, ensuring that there are no "zombie" requests
leaked due to a race between the authenticator and the pending requests.
This is guaranteed by subscription and broadcast all happening within
the authentication state lock to ensure proper synchronisation.