Simplify Channel to use an atomic array of subcribers
The critical section in `Channel.swift` involves 1) enqueueing a
subscriber and 2) broadcasting a message to all current subscribers.

Instead of dispatching multiple subscriber blocks and having threads
block waiting to receive a message, we can instead keep a list of
subscriber functions, wrapped in a mutex, and wait until broadcast time
to grab all subcriber functions and dispatch a block to execute
receiving the message.

This fixes an intermittent crash resulting from a race condition between
the broadcaster and subscribers: occasionally the broadcaster would set
`message` to nil before the signalled subscriber has woken up and read
the message, resulting in a crash when force-unwrapping a nil.