Extract TokenStorage protocol
Updates `RequestAuthorizer` to accept a value conforming to
`TokenStorage`, currently defaulting to a simple in-memory token store
(`SimpleTokenStorage`).

Removes the `service` parameter to `fetchToken()`, which is specific to
`KeychainTokenStorage`. The service name is instead injected into the
`KeychainTokenStorage` initializer. This means that a single
`KeychainTokenStorage` instance can only save a token for a single
service, however this is already how `RequestAuthorizer` is expected to
be used -- a one-to-one relationship between `RequestAuthorizer`,
`FinchProvider` and the token storage.