Add a README and CONTRIBUTING guide
The CONTRIBUTING guide is the standard thoughtbot one.

The README attempts to explain what Finch is, when you would not use it,
how to use it, how to install it, and standard thoughtbot addendum.

Only one example is provided for now, with the expectation that further
tutorials and blog posts will help for other use cases.