Extract GitHubAPIClient and update README example accordingly
The example in the README wasn't very well factored, dumping a bunch of
configuration into the app delegate. This change extracts a
`GitHubAPIClient`, with two singleton instances: one configured for
OAuth, and one configured for Basic Auth. This simplifies a lot of
duplicated demo app code, and provides a clearer example for the README.