Don't build SuperbDemo from the shared scheme
For convenience, Super.xcscheme was configured to build SuperbDemo in
the Run action. Unfortunately this also caused SuperbDemo to be built
unintentionally when building as a Carthage framework. The result is
that `bin/validate-secrets` would run, failing the build because secrets
are not set up in the Carthage environment.

This change does two things: first, it removes SuperbDemo from the
shared Superb scheme for building, restricting it to only build when
running the tests. This means that during development the SuperbDemo
target is not available for running, so we re-enable the "Autocreate
Schemes" option in Xcode so that a SuperbDemo scheme is created for
development purposes.