Replace Channel abstraction with simple request queue
The Channel abstraction, in an attempt to be generic and thread-safe,
uses synchronisation primitives to ensure safe access to its internal
subscribers array. In practice, this extra coordination overhead is
unnecessary, as `RequestAuthorizer` can instead use a private variable
and updated it safely by only ever accessing it on `messageQueue`. This
makes `RequestAuthorizer` much easier to understand, and lets us delete
a bunch of code. I haven't measured this, but it's also potentially
faster now that there are no blocking calls.