Allow requests to be cancelled
This updates `RequestAuthorizer.performAuthorized()` to return an
`AuthorizedRequestTask` object, analogous to `URLSessionDataTask`.
Calling `AuthorizedRequestTask.cancel()` will ask the
`RequestAuthorizer` to perform the cancellation.

`RequestAuthorizer` now keeps track of the state of its requests so that
it can respond to cancellation appropriately. If the request is running,
the message is essentially forward to the underlying `URLSessionDataTask`.
If the request is pending, it is marked as cancelled, and the completion
handler is called with `URLError.cancelled`, so that it is
indistinguishable from cancelling a running request.